Add login context and protect users apis

+++ backend/src/Services/Api/UsersService.php
@@ -6,16 +6,19 @@ use App\Repository\UsersRepository;
 use App\Entity\Users;
 use App\Enum\Role;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasher;
 
 class UsersService
 {
   public function __construct(
     private EntityManagerInterface $entityManager,
-    private UsersRepository $usersRepository
+    private UsersRepository $usersRepository,
+    private UserPasswordHasher $passwordHasher
   ) {}
 
-  public function getUsers()
+  public function getUsers($currentUser)
   {
+    if ($currentUser->getRole()->value !== 'ROLE_ADMIN') return 'Forbidden';
     $users = $this->usersRepository->findAll();
     if (!$users) return;
     $usersData = array_map(fn($u) => [
@@ -26,8 +29,9 @@ class UsersService
     return $usersData;
   }
 
-  public function getUser(int $id)
+  public function getUser(int $id, $currentUser)
   {
+    if ($currentUser->getRole()->value !== 'ROLE_ADMIN' && $currentUser->getId() !== (int)$id) return 'Forbidden';
     $user = $this->usersRepository->find($id);
     if (!$user) return;
     $userData = [
@@ -40,10 +44,14 @@ class UsersService
 
   public function addUser($data)
   {
+    $user = $this->usersRepository->findOneBy(['email' => $data['email']]);
+    if (!$user) return;
     $user = new Users();
     $user->setName($data['name']);
     $user->setEmail($data['email']);
     $user->setRole(Role::from($data['role']));
+    $hashedPassword = $this->passwordHasher->hashPassword($user, $data['password']);
+    $user->setPassword($hashedPassword);
 
     $this->entityManager->persist($user);
     $this->entityManager->flush();
@@ -51,10 +59,14 @@ class UsersService
     return $user;
   }
 
-  public function setUser($data, int $id)
+
+  public function setUser($data, int $id, $currentUser)
   {
     $user = $this->usersRepository->find($id);
     if (!$user) return;
+
+    if ($currentUser->getRole()->value !== 'ROLE_ADMIN' && $currentUser->getId() !== (int)$id) return 'Forbidden';
+
     $user->setName($data['name']);
     $user->setEmail($data['email']);
     $user->setRole(Role::from($data['role']));
@@ -62,10 +74,13 @@ class UsersService
     return $user;
   }
 
-  public function deleteUser(int $id)
+  public function deleteUser(int $id, $currentUser)
   {
     $user = $this->usersRepository->find($id);
     if (!$user) return;
+
+    if ($currentUser->getRole()->value !== 'ROLE_ADMIN' && $currentUser->getId() !== (int)$id) return 'Forbidden';
+
     $this->entityManager->remove($user);
     $this->entityManager->flush();
